Rename old storage
Ready for transition to SQLite
