+++ ledger_bot/reminder_manager.py
@@ -11,7 +11,7 @@ from apscheduler.schedulers.asyncio import AsyncIOScheduler
 
 from .message_generators import generate_reminder_status_message
 from .models import BotMessage, Reminder, Transaction
-from .storage import AirtableStorage
+from .storage_airtable import AirtableStorage
 
 if TYPE_CHECKING:
     from .LedgerBot import LedgerBot
