+++ ledger_bot/scheduled_commands/cleanup.py
@@ -7,7 +7,7 @@ import discord
 
 from ledger_bot.errors import AirTableError
 from ledger_bot.models import BotMessage, Transaction
-from ledger_bot.storage import AirtableStorage
+from ledger_bot.storage_airtable import AirtableStorage
 
 if TYPE_CHECKING:
     from ledger_bot.LedgerBot import LedgerBot
