+++ ledger_bot/slash_commands/command_help.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any, Dict
 import discord
 
 from ledger_bot.message_generators import generate_help_message
-from ledger_bot.storage import AirtableStorage
+from ledger_bot.storage_airtable import AirtableStorage
 
 if TYPE_CHECKING:
     from ledger_bot.LedgerBot import LedgerBot
