+++ ledger_bot/slash_commands/command_new_sale.py
@@ -10,7 +10,7 @@ from ledger_bot.errors import AirTableError
 from ledger_bot.LedgerBot import LedgerBot
 from ledger_bot.message_generators import generate_transaction_status_message
 from ledger_bot.models import Transaction
-from ledger_bot.storage import AirtableStorage
+from ledger_bot.storage_airtable import AirtableStorage
 
 log = logging.getLogger(__name__)
